Rename module async to async_util
